Add dynamic plural kind lookup into reconcile step
If not user specified, then at first attempt the plural for a kind will
be looked up within the resource function's first reconcile. This should
make it possible to omit the plural form and instead allow it to be
dynamically looked up. This is less efficient, but much more ergonomic
for developers.