Handle create contention
If there are multiple parallel attempts to create the same resource, all
but one will fail with 409s. This adjusts the prior behavior to return a
retry from resource function as though the create occurred.

Note, other error codes should be reviewed to determine if they should
retry as well.